[FE] modify: 마이페이지 디데이 목록 로그인과 동시에 state 저장
