+++ FE/src/pages/MyPage.tsx
@@ -10,7 +10,7 @@ import MyFriend from "@components/mypage/details/friend/MyFriend";
 import { settingApi } from "@api/Api";
 import { useAppDispatch, useAppSelector } from "@hooks/hook";
 import { selectUserId } from "@store/authSlice";
-import { setCategoryColors, setCategoryInput, setCategoryList, setDdayList } from "@store/mypageSlice";
+import { setCategoryColors, setCategoryInput, setCategoryList } from "@store/mypageSlice";
 import { CategoryConfig } from "@util/planner.interface";
 
 const MyPage = () => {
@@ -43,20 +43,9 @@ const MyPage = () => {
       });
   };
 
-  const getDdayList = () => {
-    settingApi
-      .ddays(userId)
-      .then((res) => {
-        const response = res.data.data.ddayList;
-        dispatch(setDdayList(response));
-      })
-      .catch((err) => console.log(err));
-  };
-
   useEffect(() => {
     getCategoryList();
     getCategoryColors();
-    getDdayList();
   }, []);
 
   return (
